requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
SQLAlchemy==1.2.12
flask-sqlalchemy==2.3.2
python-decouple==3.1
Flask-Migrate==2.2.1
Mako==1.0.7
alembic==1.0.0
python-dateutil==2.7.3
python-editor==1.0.3
